Accept and delete logic applied

+++ app/(components)/DeleteBlock.jsx
@@ -6,9 +6,12 @@ import React from 'react'
 import  {useRouter } from 'next/navigation'
 
 const  DeleteBlock = ({id}) => {
-
     const router = useRouter()
     const deleteContract = async () =>{
+        const confirmation =   confirm("Are you sure want to delete?");
+        if(!confirmation){
+            return
+        }
         const res = await fetch(`http://localhost:3000/api/Contracts/${id}`, {
           method: "DELETE",
         });
@@ -16,8 +19,6 @@ const  DeleteBlock = ({id}) => {
         if(res.ok) {
             router.refresh()
         }
-
-
     }
   return (
     <FontAwesomeIcon  icon={faTrash} className='text-red-400 hover:cursor-pointer hover:text-red-200' onClick={deleteContract} />
